Move issue report text to language files.
